fixed cast

+++ Quadratic.java
@@ -1,5 +1,3 @@
-import javax.naming.NameNotFoundException;
-
 public class Quadratic {
     public static void main(String[] args) {
         int a = Integer.parseInt(args[0]);
@@ -8,11 +6,11 @@ public class Quadratic {
         double bsq = Math.pow(b, 2);
       
         double discrim = (bsq - (4 * a * c));
-         double result1 =  (-b + Math.sqrt(bsq - discrim)) / (2 * a);
+        double result1 =  (-b + Math.sqrt(bsq - discrim)) / (2 * a);
         double result2 =  (-b - Math.sqrt(bsq - discrim)) / (2 * a);
        
         if (a == 0) {
-            double noa = (-c / b);
+            double noa = -c / (double) b;
             System.out.println(noa);
         } else if (discrim < 0) {
             System.out.println("no solution");
